Extract network utils to separate module
