+++ winnow/feature_extraction/utils.py
@@ -2,24 +2,7 @@ import os
 
 import cv2
 import numpy as np
-import requests
 import yaml
-
-
-def download_file(local_filename, url):
-
-    with requests.get(url, stream=True) as r:
-
-        r.raise_for_status()
-        with open(local_filename, "wb") as f:
-
-            for chunk in r.iter_content(chunk_size=8192):
-
-                if chunk:
-
-                    f.write(chunk)
-
-        return local_filename
 
 
 def load_video(video, desired_size, frame_sampling):
